Task2/requirements.txt
streamlit>=1.29.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
gspread>=5.12.0